Developer desarrollo

+++ s2_sitiosweb-api/src/main/java/co/edu/uniandes/csw/sitiosweb/dtos/UserDTO.java
@@ -16,7 +16,7 @@ public class UserDTO implements Serializable {
     private Long id;
     private String login;
     private String email;
-    private Long phone;
+    private Integer phone;
     
     /**
      * Constructor por defecto
@@ -98,14 +98,14 @@ public class UserDTO implements Serializable {
     /**
      * @return the phone
      */
-    public Long getPhone() {
+    public Integer getPhone() {
         return phone;
     }
 
     /**
      * @param phone the phone to set
      */
-    public void setPhone(Long phone) {
+    public void setPhone(Integer phone) {
         this.phone = phone;
     }
